+++ test/read_mixer.c
@@ -115,7 +115,7 @@ filter_plugin_by_name(G_GNUC_UNUSED const char *name)
 }
 
 bool
-pcm_volume(G_GNUC_UNUSED void *buffer, G_GNUC_UNUSED int length,
+pcm_volume(G_GNUC_UNUSED void *buffer, G_GNUC_UNUSED size_t length,
 	   G_GNUC_UNUSED enum sample_format format,
 	   G_GNUC_UNUSED int volume)
 {
